Revert "arm, arm926ejs: Flush the data cache before disabling it"
This reverts commit da104e04ecc30b90fdc121737f220f0d9c252196.

Picoxcell devices contain an ARM1176 core, however for historical
reasons we specify an ARM926 core in the U-Boot build.
The change that was introduced in commit da104e... introduced a
co-processor instruction that is an underfind instruction in an ARM1176.
The easiest fix to to just revert the change.

This commit forms part of the resolution of Mantis #10482