picoxcell: compile with -march=armv5
This commit changes the architecture defined when building u-boot
for ARM926 targets. Originally this was defined as armv5te. Using
armv5te allows the compiler to use ldrd and strd instructions.
This causes problems (at best data aborts, at worst incorrect data
is used) with non aligned data (specifically some variables used in
the picoxcell timer implementation that are contained in the
global_data structure).

The other fix is to align the pointer used for the global_data
struct. Go check out the board_init_f() in arch/arm/lib/board.c

This commit forms the resoultion of Mantis Issue #10482 and
Mantis Issue #10530