Implement generalised public exponents for verified boot
The initial verified boot implementation was limited to a public
exponent of 65537 (F4), which makes the verified boot
implementation difficult to use with an existing PKI
infrastructure that uses different public exponent values.  This
is especially true of an existing infrastructure based on
hardware security modules, where there may be no available HSM
slots for additional keys.

This patch makes changes to record the public exponent in the
configuration OF tree and to use this public exponent in
verification.  When no public exponent is found in the config OF
tree the code falls back to the original default value (65537)
to maintain compatibility with the initial implementation.

The mathematics and implementation for this change was done by
Andrew Bott <Andrew.Bott@ipaccess.com>, I just cleaned the
implementation up for mainline inclusion and wrote the changes
to rsa-sign.c.

Signed-off-by: Andrew Bott <Andrew.Bott@ipaccess.com>