PHY address fallback and error return
The existing eTSEC driver does not recognise PHY connect failures.

This change recognises connect failures and falls back to a
different well known address to keep U-Boot working on a difference
between two IPA boards.  This change also returns an error from
the init routine if both PHY connect commands fail, which did not
happen before.

The PHY address used for each of the eTSEC interfaces on the board
is stored in an environment variable, ipa_phyaddr_<eTSEC_NAME>.

Signed-off-by: Michael van der Westhuizen <michael@smart-africa.com>