Revocation for ml9131 and ipa9131
This changeset introduces revocation bit counting and value
construction for loader and application revocation.  The changeset
then builds on this functionality to do self revocation of the
ml9131 microloader, perform loader revocation checks of the load
of U-Boot from an IBN image (with corrections) and perform
fitImage revocation checks against the application revocation
from normal U-Boot.

Signed-off-by: Michael van der Westhuizen <michael@smart-africa.com>